# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Skoll API
  version: 1.0.0
  description: OpenAPI document for Skoll core HTTP endpoints.
servers:
  - url: __API_BASE_PREFIX__
paths:
...
  /v1/plugins/pharma_oa/api/products:
    get:
      summary: List Pharma OA products
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, disabled]
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: cursor
          in: query
          description: Opaque cursor returned by the previous page. When present, it takes precedence over offset.
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PharmaMasterListAPIResponse'}
    post:
      summary: Create Pharma OA product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name, spec, dosageForm, manufacturer, approvalNumber]
              properties:
                code:
                  type: string
                name:
                  type: string
                spec:
                  type: string
                dosageForm:
                  type: string
                manufacturer:
                  type: string
                approvalNumber:
                  type: string
                temperature:
                  type: object
                  properties:
                    required:
                      type: boolean
                    minCelsius:
                      type: number
                    maxCelsius:
                      type: number
      responses:
        '201':
          description: Created
  /v1/plugins/pharma_oa/api/products/{id}:
    put:
      summary: Update Pharma OA product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /v1/plugins/pharma_oa/api/products/{id}/disable:
    post:
      summary: Disable Pharma OA product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /v1/plugins/pharma_oa/api/products/import:
    post:
      summary: Import Pharma OA products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rows:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: OK
...
  /v1/plugins/pharma_oa/api/purchase-requests:
    get:
      summary: List Pharma OA purchase requests
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestListResponse'
    post:
      summary: Create a Pharma OA purchase request and start approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequestCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestResponse'
        '400':
          description: Invalid request or supplier is not purchase eligible
  /v1/plugins/pharma_oa/api/purchase-requests/{id}:
    get:
      summary: Get a Pharma OA purchase request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestResponse'
        '404':
          description: Purchase request not found
  /v1/plugins/pharma_oa/api/purchase-requests/{id}/approve:
    post:
      summary: Approve a purchase request and generate its purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseApprovalRequest'
      responses:
        '200':
          description: Approved purchase order; repeated approval returns the same order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrderResponse'
        '400':
          description: Approval failed or supplier is not purchase eligible
        '403':
          description: Forbidden
  /v1/plugins/pharma_oa/api/purchase-requests/{id}/reject:
    post:
      summary: Reject a purchase request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseApprovalRequest'
      responses:
        '200':
          description: Rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseRequestResponse'
        '403':
          description: Forbidden
...
  /v1/plugins/pharma_oa/api/purchase-orders/{id}:
    get:
      summary: Get a Pharma OA purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrderResponse'
        '404':
          description: Purchase order not found
  /v1/plugins/pharma_oa/api/purchase-inbounds:
    get:
      summary: List Pharma OA purchase inbounds
      responses:
        '200': {description: OK}
        '403': {description: Forbidden}
    post:
      summary: Complete purchase inbound and increase batch inventory
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PurchaseInboundCreateRequest'}
      responses:
        '201': {description: Created}
        '400': {description: Invalid order, location, batch, expiry, attachment, or quantity}
        '403': {description: Forbidden}
  /v1/plugins/pharma_oa/api/purchase-inbounds/{id}:
    get:
      summary: Get Pharma OA purchase inbound
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: OK}
        '403': {description: Forbidden}
        '404': {description: Purchase inbound not found}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Skoll JWT access token. Aggregated plugin API operations require this scheme and declare x-skoll-permission.
  schemas:
...
    PharmaMasterListData:
      type: object
      required: [items, offset, limit, total, hasMore, nextCursor, sort]
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        offset: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 200}
        total: {type: integer, format: int64, minimum: 0}
        hasMore: {type: boolean}
        nextCursor:
          type: string
          description: Empty on the final page; otherwise pass this value as the next request cursor.
        sort:
          type: string
          enum: ['code:asc']
    PharmaMasterListAPIResponse:
      allOf:
        - $ref: '#/components/schemas/MessageResponse'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/PharmaMasterListData'
...
    PurchaseInboundAttachment:
      type: object
      required: [fileId, fileName, size]
      properties:
        fileId: {type: string}
        fileName: {type: string}
        size: {type: integer, format: int64, minimum: 0}
    PurchaseInboundLine:
      type: object
      required: [productId, quantity, batchNo, productionDate, expiresAt]
      properties:
        productId: {type: string}
        quantity: {type: integer, minimum: 1}
        batchNo: {type: string}
        productionDate: {type: string, format: date-time}
        expiresAt: {type: string, format: date-time}
        batchId: {type: string, readOnly: true}
        ledgerId: {type: string, readOnly: true}
    PurchaseInboundCreateRequest:
      type: object
      required: [number, purchaseOrderId, warehouseId, areaId, locationId, lines, actorId]
      properties:
        number: {type: string}
        purchaseOrderId: {type: string}
        warehouseId: {type: string}
        areaId: {type: string}
        locationId: {type: string}
        actorId: {type: string}
        lines: {type: array, minItems: 1, items: {$ref: '#/components/schemas/PurchaseInboundLine'}}
        attachments: {type: array, items: {$ref: '#/components/schemas/PurchaseInboundAttachment'}}
    PurchaseLine:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:
          type: string
        quantity:
          type: number
          format: double
          exclusiveMinimum: 0
        unitPrice:
          type: number
          format: double
          minimum: 0
        amount:
          type: number
          format: double
          readOnly: true
    PurchaseRequestCreateRequest:
      type: object
      required: [number, supplierId, requesterId, approverId, lines]
      properties:
        number: {type: string}
        supplierId: {type: string}
        requesterId: {type: string}
        approverId: {type: string}
        reason: {type: string}
        lines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PurchaseLine'
    PurchaseApprovalRequest:
      type: object
      required: [actorId]
      properties:
        actorId: {type: string}
        comment: {type: string}
    PurchaseRequest:
      type: object
      required: [id, number, supplierId, requesterId, approverId, lines, totalAmount, status, workflowInstanceId]
      properties:
        id: {type: string}
        number: {type: string}
        supplierId: {type: string}
        requesterId: {type: string}
        approverId: {type: string}
        reason: {type: string}
        lines:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseLine'
        totalAmount: {type: number, format: double}
        status: {type: string, enum: [pending_approval, approved, rejected]}
        workflowInstanceId: {type: string}
        purchaseOrderId: {type: string}
    PurchaseOrder:
      type: object
      required: [id, number, purchaseRequestId, supplierId, lines, totalAmount, status, approvedBy, approvedAt]
      properties:
        id: {type: string}
        number: {type: string}
        purchaseRequestId: {type: string}
        supplierId: {type: string}
        lines:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseLine'
        totalAmount: {type: number, format: double}
        status: {type: string, enum: [open]}
        approvedBy: {type: string}
        approvedAt: {type: string, format: date-time}
    PurchaseRequestResponse:
      type: object
      required: [code, message, data]
      properties:
        code: {type: string}
        message: {type: string}
        data:
          type: object
          required: [item]
          properties:
            item: {$ref: '#/components/schemas/PurchaseRequest'}
    PurchaseRequestListResponse:
      type: object
      required: [code, message, data]
      properties:
        code: {type: string}
        message: {type: string}
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: {$ref: '#/components/schemas/PurchaseRequest'}
    PurchaseOrderResponse:
      type: object
      required: [code, message, data]
      properties:
        code: {type: string}
        message: {type: string}
        data:
          type: object
          required: [item]
          properties:
            item: {$ref: '#/components/schemas/PurchaseOrder'}
...
    MessageResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
...